pyproject.toml
[tool.poetry]
name = "shift-service"
version = "0.1.0"
description = ""
authors = ["Evgeniy Berezovskiy <geoche74@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = ">=3.0.0,<3.1.0"
fastapi = ">=0.109.0,<0.110.0"
pydantic = ">=2.5.3,<2.6.0"
sqlalchemy = ">=2.0.25,<2.1.0"
uvicorn = ">=0.25.0,<0.26.0"
alembic = ">=1.13.1,<1.14.0"
asyncpg = ">=0.29.0,<0.30.0"
python-multipart = ">=0.0.6,<0.1.0"
gunicorn = ">=21.2.0,<21.3.0"
fastapi-filter = ">=1.1.0,<1.2.0"
httpx = ">=0.27.0,<0.28.0"
dependency-injector = ">=4.41.0,<4.42.0"
pydantic-settings = "^2.2.1"
opentelemetry-distro = "^0.45b0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-sqlalchemy = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-jaeger = "^1.21.0"
apscheduler = "^3.10.4"
testcontainers = "^4.4.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
pre-commit = "^3.7.0"
ruff = "^0.3.5"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0,<8.1.0"
pytest-alembic = "^0.11.0"
pytest-mock-resources = "^2.10.3"
python-on-whales = "^0.70.1"
pytest-asyncio = "^0.23.7"
async-asgi-testclient = "^1.4.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
strict_equality = true
implicit_reexport = false
show_error_context = true
pretty = true
show_traceback = true
scripts_are_modules = true
ignore_missing_imports = true
namespace_packages = true
exclude = ["alembic", "tests",]

[tool.ruff]
line-length = 88
exclude = ["tests"]
lint.select = ["I", "ASYNC", "S", "T20", "SIM", "E", "W", "C90", "PL", "RUF", "F"]
lint.ignore = ["RUF001", "RUF002", "RUF012"]
lint.pylint.max-args = 8

[tool.pytest.ini_options]
filterwarnings = ["ignore::pydantic.PydanticDeprecatedSince20", ]